requirements.txt
Django==5.2.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
djoser==2.2.3
drf-yasg==1.21.7
PyYAML==6.0.1
uritemplate==4.1.1